@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/services.dart';
 import 'package:flutter_facebook_login/flutter_facebook_login.dart';
 import 'package:flutter_facebook_login/src/clock.dart';
@@ -41,9 +43,9 @@ void main() {
     };
 
     final log = <MethodCall>[];
-    FacebookLogin sut;
+    late FacebookLogin sut;
 
-    void setMethodCallResponse(Map<String, dynamic> response) {
+    void setMethodCallResponse(Map<String, dynamic>? response) {
       channel.setMockMethodCallHandler((methodCall) {
         log.add(methodCall);
         return Future.value(response);
@@ -94,7 +96,7 @@ void main() {
       setMethodCallResponse(kLoggedInResponse);
 
       final result = await sut.logIn([]);
-      final map = result.accessToken.toMap();
+      final map = result.accessToken!.toMap();
 
       expect(
         map,
@@ -122,13 +124,6 @@ void main() {
       final second = FacebookAccessToken.fromMap(kAccessToken);
 
       expect(first, equals(second));
-    });
-
-    test('loginBehavior - with null argument', () async {
-      setMethodCallResponse(null);
-
-      // Setting a null login behavior is not allowed.
-      expect(() => sut.loginBehavior = null, throwsAssertionError);
     });
 
     test('loginBehavior - nativeWithFallback is the default', () async {
@@ -187,7 +182,7 @@ void main() {
       ]);
 
       expect(result.status, FacebookLoginStatus.loggedIn);
-      expectAccessTokenParsedCorrectly(result.accessToken);
+      expectAccessTokenParsedCorrectly(result.accessToken!);
 
       expect(
         log,
@@ -278,7 +273,7 @@ void main() {
         () async {
       setMethodCallResponse(kAccessToken);
 
-      final accessToken = await sut.currentAccessToken;
+      final accessToken = await (sut.currentAccessToken as FutureOr<FacebookAccessToken>);
       expectAccessTokenParsedCorrectly(accessToken);
     });
 
@@ -288,7 +283,7 @@ void main() {
 
       Clock.dateTimeResolver = () => beforeExpiry;
 
-      final accessToken = await sut.currentAccessToken;
+      final accessToken = await (sut.currentAccessToken as FutureOr<FacebookAccessToken>);
       expect(accessToken.isValid(), isTrue);
     });
 
@@ -298,7 +293,7 @@ void main() {
 
       Clock.dateTimeResolver = () => afterExpiry;
 
-      final accessToken = await sut.currentAccessToken;
+      final accessToken = await (sut.currentAccessToken as FutureOr<FacebookAccessToken>);
       expect(accessToken.isValid(), isFalse);
     });
   });
